refactor: Call helper function
The assertion assert_coin_select() was independently calling the search
algo and then asserting the results.  Reduce code complexity and line
count by simply calling assert_coin_select_params().

Remove the FEE_RATE const since now the fee_rate for the test is used in
one place when creating the params to be passed to
asswert_coin_select_params().

Remove `build_pool()` for the same reason as removing FEE_RATE const
since this is defined in the params now.