selection: move parameters from function call into `UTXO` field
When enforcing a `max_weight` (TK), the `UTXO` Ordering needs to be on
effective_value.  However, effective_value isn't a property of `UTXO`
and is calculated per UTXO by each of the selection algorithms.
Therefore, this commit kills two birds by making the code more dry such
that the effective_value calculations are removed from each individual
selection algorithms.  Furthermore, by adding effective_value as a
property of `UTXO`, the order can be enforced.

To add `effective_value` to the `UTXO`, it is calculated during
construction.  Therefore, `fee_rate` and `long_term_fee_rate` are also
needed in the constructor.  Therefore, these parameters are removed from
being passed to each individual algorithms and instead passed to the
constructor of the `UTXO`.  This also has another pleasant side-effect
in that it matches how bitcoin-core works having in that the `fee_rate`
and `long_term_fee_rate` can be used to store the `fee` and
`long_term_fee` of the `UTXO`.

Since `fee_rate`, `long_term_fee_rate` and `effective_value` all work
as a set of constraints on coin-selection, this change was done in this
single commit.  Lastly, by adding these parameters to the `UTXO`
constructor, this makes it impossible to create a `UTXO` which has a
negative effective value.  This simplifies the logic of the selection
algorithms and also invalidates some tests that assert the behavior of
handling negative valued `UTXOs`.

A marginal performance improvment for BnB and a large imprvement for
`SRD` is recorded.  The large improvement in `SRD` reflects that the
`effective_value` is not recalculated in loop and also there is no
longer a needed check that `effecive_value` is a positive value within
bounds.